refactoring : user merge 수정
