+++ README.md
@@ -1,2 +1,2 @@
-# umchiumtee_back
-음치음티 백엔드 레포지토리
+# lalalia_back
+랄라리아 백엔드 레포지토리
