+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'umchiumtee'
+rootProject.name = 'lalalia'
