--- src/main/java/com/isthisteamisthis/umchiumtee/user/command/application/dto/request/SignUpRequestDTO.java
@@ -1,17 +1,0 @@
-package com.isthisteamisthis.umchiumtee.user.command.application.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Setter
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignUpRequestDTO {
-
-    private String userId;
-    private String password;
-
-}
