+++ src/main/resources/db/data/data.sql
@@ -1,2 +1,2 @@
 INSERT INTO TBL_USER(user_no, user_id, user_intro, email, category)
-VALUES (1, '가수지망생', '가수가 되고 싶은 학생입니다!', 'singer@gmail.com', 'SINGER');
+VALUES (1, 1234123530, '가수가 되고 싶은 학생입니다!', 'singer@gmail.com', 'SINGER');
